+++ DKPlugins/DK/DKUtil.h
@@ -60,12 +60,12 @@ public:
 	static bool KeyIsDown(int& key);
 
 	//https://stackoverflow.com/questions/63166/how-to-determine-cpu-and-memory-consumption-from-inside-a-process
-	static bool VirtualMemory(int& virtualMemory);
-	static bool VirtualMemoryUsed(int& virtualMemory);
-	static bool VirtualMemoryUsedByApp(int& virtualMemory);
-	static bool PhysicalMemory(int& physicalMemory);
-	static bool PhysicalMemoryUsed(int& physicalMemory);
-	static bool PhysicalMemoryUsedByApp(int& physicalMemory);
+	static bool VirtualMemory(float& virtualMemory);
+	static bool VirtualMemoryUsed(float& virtualMemory);
+	static bool VirtualMemoryUsedByApp(float& virtualMemory);
+	static bool PhysicalMemory(float& physicalMemory);
+	static bool PhysicalMemoryUsed(float& physicalMemory);
+	static bool PhysicalMemoryUsedByApp(float& physicalMemory);
 	static bool CpuUsed(int& cpu);
 	static bool CpuUsedByApp(int& cpu);
